backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
supabase==2.0.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.0
supabase==2.0.3
faiss-cpu==1.9.0.post1
scikit-learn
groq
numpy
pydantic
aiofiles==23.2.1
